updating with documentation, moving WIP pieces into tests
