+++ Sources/CASimEngine/Documentation.docc/CASimulationStorage.md
@@ -1,0 +1,16 @@
+# ``CASimulationStorage``
+
+## Topics
+
+### Creating Storage
+
+- ``init(_:)``
+- ``T``
+
+### Inspecting Storage
+
+- ``bounds``
+
+### Retrieving Updates
+
+- ``changes()``
