+++ Sources/CASimEngine/Documentation.docc/SwapStep.md
@@ -1,0 +1,8 @@
+# ``SwapStep``
+
+## Topics
+
+### Defining Simulation Rules
+
+- ``perform(storage0:storage1:)``
+- ``StorageType``
