+++ Tests/CASimEngineTests/Comparable+clamped.swift
