+++ Tests/CASimEngineTests/Resource+examples.swift
